Revert "riscv: mm: support Svnapot in huge vmap"
This reverts commit ce173474cf19fe7fbe8f0fc74e3c81ec9c3d9807.

We cannot (for now) correctly deal with NAPOT mappings in vmalloc/vmap
because if some part of a NAPOT mapping is unmapped, the logic to break the
NAPOT mapping does not exist yet. For example:

ptr = vmalloc_huge(64 * 1024, GFP_KERNEL);
vunmap_range((unsigned long)(ptr + PAGE_SIZE),
	     (unsigned long)(ptr + 64 * 1024));

leads to the following kernel page table dump:

0xffff8f8000ef0000-0xffff8f8000ef1000    0x00000001033c0000         4K PTE N   ..     ..   D A G . . W R V

Meaning the first entry which was not unmapped still has the N bit set,
which, if accessed first and cached in the TLB, could allow access to the
unmapped range.

So fix this by reverting the commit that introduced the vmap/vmalloc
support.

Signed-off-by: Alexandre Ghiti <alexghiti@rivosinc.com>