RIVOS: rasd: Handle inability to get all MSIs
Some interrupt controllers do not support more than one MSI vector for a
device.

Get probe to still work in this case.

Further work is required to communicate this state of affairs to the
accelerator sw so that the available vector is shared.
We should probably keep track of this on the host driver sidei as well.