DO NOT UPSTREAM: RIVOS: mm: Support MAP_UNINITIALIZED with MMU
Support CONFIG_MMAP_ALLOW_UNINITIALIZED and the MAP_UNINITIALIZED mmap()
flag on MMU kernels so that userspace can request uninitialized memory
when it explicitly does not need the memory to be zeroed out. A VMA flag
is added to indicate which VMAs have requested uninitialized memory.

This is intended for RTL simulation/emulation use cases where the zeroing
out of memory is prohibitively expensive. It is for testing purposes only
and must NEVER BE USED IN PRODUCTION for obvious oreason.

Signed-off-by: Andrew Bresticker <abrestic@rivosinc.com>