mm: Provide addr parameter to page_table_check_pte_set()
To provide support for powerpc platforms, provide an address parameter
to the page_table_check_pte_set() routine. This parameter is needed on
some powerpc platforms which do not encode whether a mapping is for
user or kernel in the pte. On such platforms, this can be inferred
from the addr parameter.

Signed-off-by: Rohan McLure <rmclure@linux.ibm.com>
Signed-off-by: Björn Töpel <bjorn@rivosinc.com>