binfmt_flat: Fix corruption when not offsetting data start
Commit 04d82a6d0881 ("binfmt_flat: allow not offsetting data start")
introduced a RISC-V specific variant of the FLAT format which does not
allocate any space for the (obsolescent) array of shared library
pointers. However, it did not disable the code which initializes the
array, resulting in the corruption of sizeof(long) bytes before the DATA
segment, generally the end of the TEXT segment.

Use CONFIG_BINFMT_FLAT_NO_DATA_START_OFFSET to guard initialization the
shared library pointer region so that it will only be initialized if
space is reserved for it.

Fixes: 04d82a6d0881 ("binfmt_flat: allow not offsetting data start")
Signed-off-by: Stefan O'Rear <sorear@fastmail.com>
Tested-by: Waldemar Brodkorb <wbx@openadk.org>
Reviewed-by: Damien Le Moal <dlemoal
Signed-off-by: Björn Töpel <bjorn@rivosinc.com>