riscv: dts: starfive: remove non-existant spi device from jh7110-common.dtsi
There is no rohm,dh2228fv on any of supported JH7110 boards - in fact
the dh2228fv almost certainly does not exist as it is not a valid Rohm
part number. Likely a typo by Maxime when adding the device originally,
and should have been bh2228fv, but these boards do not have a bh2228fv
either! Remove it from jh7110-common.dtsi - pretending to have a device
so that the spidev driver will be bound by Linux is not acceptable.

Fixes: 74fb20c8f05d ("riscv: dts: starfive: Add spi node and pins configuration")
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Reviewed-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Reviewed-by: Emil Renner Berthing <emil.renner.berthing@canonical.com>
Signed-off-by: Björn Töpel <bjorn@rivosinc.com>