cpumask: de-duplicate assign_cpu() API
We've got cpumask_assign_cpu() function and assign_cpu() macro, both
doing the same thing. We need to drop one to avoid unneeded duplicatioon.

Now that underlying assign_bit() implemented as a macro, it would make
sense to keep assign_cpu() which is also implemented as a macro, in sake
of unification.

This patch also removes __cpumask_assign_cpu() as the function is
unused.

Signed-off-by: Yury Norov <yury.norov@gmail.com>
Signed-off-by: Björn Töpel <bjorn@rivosinc.com>