RISC-V: Schedule SBI-0.1 for removal
We've said we're going to deprecate SBI-0.1 for a while now, but I
didn't find that written down anywhere.  So mark SBI-0.1 as deprecated
and print a timeline for that deprecation on systems that probe it.

This also hides the Kconfig behind EXPERT, in the hope that doing so
will poke users to notice the deprecation.

Signed-off-by: Palmer Dabbelt <palmer@rivosinc.com>
Signed-off-by: Björn Töpel <bjorn@rivosinc.com>