RISC-V: defconfig: Remove ARCH_RENESAS
The RZ/Five is gated behind CONFIG_NONPORTABLE because of the DMA pool
issue.  e36ddf322686 ("riscv: defconfig: Disable RZ/Five peripheral
support") removes the drivers, so let's remove the rest of the vendor
support as well -- users of these will already have to play aruond with
kernel configs, so this shouldn't be much of a burden.

Signed-off-by: Palmer Dabbelt <palmer@rivosinc.com>
Signed-off-by: Björn Töpel <bjorn@rivosinc.com>