riscv: dts: starfive: add framework dts
Add framework dts to support RISC-V Framework Laptop 13 Mainboard.

DeepComputing is creating a RISC-V Framework Laptop 13 Mainboard
powered by a StarFive JH7110 processor.This is a developer-focused
product, aimed at making tinkering with RISC-V more accessible.
The Mainboard is under development and we'll share more on functionality,
compatibility, and availability as the program progresses.

For this Mainboard, the Model name in Framework is Framework FRANME0000;
the Model name in DeepComputing is DeepComputing fm7110.

The board features:

- JH7110 SoC
- 8 GiB LPDDR4 DRAM
- 1x Type-c(PD3.0:60W Max;usb3.2 Gen1;usb2.0;DP1.4:4K@30Hz/2.5K@60Hz)
- 1x Type-c(PD3.0:60W Max,usb3.2 Gen1;usb2.0)
- 2x Type-c(usb3.2 Gen1;usb2.0)
- 1x eMMC slot Support 32G/128G
- 1x MicroSD slot
- 1x QSPI Flash
- 1x PCIe Wi-Fi (Support AMD RZ616 Wi-Fi 6E,Intel Wi-Fi 6E AX210 No vPro)

Signed-off-by: Sandie Cao <sandie.cao@deepcomputing.io>
Signed-off-by: Björn Töpel <bjorn@rivosinc.com>