irqchip/riscv-intc: Fix no-SMP boot with ACPI
When CONFIG_SMP is disabled, the static array rintc_acpi_data with size
NR_CPUS will not be sufficient to hold all RINTC structures passed from
the firmware. All RINTC structures are required to configure
IMSIC/APLIC/PLIC properly irrespective of SMP in the OS. So, allocate
dynamic memory based on the number of RINTC structures in MADT to fix
this issue.

Fixes: f8619b66bdb1 ("irqchip/riscv-intc: Add ACPI support for AIA")
Reported-by: Björn Töpel <bjorn@kernel.org>
Closes: https://github.com/linux-riscv/linux-riscv/actions/runs/11280997511/job/31375229012
Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>
Signed-off-by: Björn Töpel <bjorn@rivosinc.com>