mm: Add macros ahead of supporting boot-time page size selection
arm64 can support multiple base page sizes. Instead of selecting a page
size at compile time, as is done today, we will make it possible to
select the desired page size on the command line.

In this case PAGE_SHIFT and it's derivatives, PAGE_SIZE and PAGE_MASK
(as well as a number of other macros related to or derived from
PAGE_SHIFT, but I'm not worrying about those yet), are no longer
compile-time constants. So the code base needs to cope with that.

As a first step, introduce MIN and MAX variants of these macros, which
express the range of possible page sizes. These are always compile-time
constants and can be used in many places where PAGE_[SHIFT|SIZE|MASK]
were previously used where a compile-time constant is required.
(Subsequent patches will do that conversion work). When the arch/build
doesn't support boot-time page size selection, the MIN and MAX variants
are equal and everything resolves as it did previously.

Additionally, introduce DEFINE_GLOBAL_PAGE_SIZE_VAR[_CONST]() which wrap
global variable defintions so that for boot-time page size selection
builds, the variable being wrapped is initialized at boot-time, instead
of compile-time. This is done by defining a function to do the
assignment, which has the "constructor" attribute. Constructor is
preferred over initcall, because when compiling a module, the module is
limited to a single initcall but constructors are unlimited. For
built-in code, constructors are now called earlier to guarrantee that
the variables are initialized by the time they are used. Any arch that
wants to enable boot-time page size selection will need to select
CONFIG_CONSTRUCTORS.

These new macros need to be available anywhere PAGE_SHIFT and friends
are available. Those are defined via asm/page.h (although some arches
have a sub-include that defines them). Unfortunately there is no
reliable asm-generic header we can easily piggy-back on, so let's define
a new one, pgtable-geometry.h, which we include near where each arch
defines PAGE_SHIFT. Ugh.

-------

Most of the problems that need to be solved over the next few patches
fall into these broad categories, which are all solved with the help of
these new macros:

1. Assignment of values derived from PAGE_SIZE in global variables

  For boot-time page size builds, we must defer the initialization of
  these variables until boot-time, when the page size is known. See
  DEFINE_GLOBAL_PAGE_SIZE_VAR[_CONST]() as described above.

2. Define static storage in units related to PAGE_SIZE

  This static storage will be defined according to PAGE_SIZE_MAX.

3. Define size of struct so that it is related to PAGE_SIZE

  The struct often contains an array that is sized to fill the page. In
  this case, use a flexible array with dynamic allocation. In other
  cases, the struct fits exactly over a page, which is a header (e.g.
  swap file header). In this case, remove the padding, and manually
  determine the struct pointer within the page.

4. BUILD_BUG_ON() with values derived from PAGE_SIZE

  In most cases, we can change these to compare againt the appropriate
  limit (either MIN or MAX). In other cases, we must change these to
  run-time BUG_ON().

5. Ensure page alignment of static data structures

  Align instead to PAGE_SIZE_MAX.

6. #ifdeffery based on PAGE_SIZE

  Often these can be changed to c code constructs. e.g. a macro that
  returns a different value depending on page size can be changed to use
  the ternary operator and the compiler will dead code strip it for the
  compile-time constant case and runtime evaluate it for the non-const
  case. Or #if/#else/#endif within a function can be converted to c
  if/else blocks, which are also dead code stripped for the const case.
  Sometimes we can change the c-preprocessor logic to use the
  appropriate MIN/MAX limit.

Signed-off-by: Ryan Roberts <ryan.roberts@arm.com>
Signed-off-by: Björn Töpel <bjorn@rivosinc.com>